chore: Move notification helpers
