+++ projects/Mallard/src/push-notifications/notification-tracking.ts
@@ -1,0 +1,11 @@
+import { notificationTrackingUrl } from 'src/helpers/settings/defaults'
+
+const notificationTracking = (
+    notificationId: string,
+    type: 'received' | 'downloaded',
+) => {
+    const url = notificationTrackingUrl(notificationId, type)
+    return fetch(url)
+}
+
+export { notificationTracking }
